+++ gatsby-browser.js
@@ -1,0 +1,7 @@
+import './src/styles/tailwind.css';
+// custom typefaces
+import '@fontsource/montserrat/variable.css';
+import '@fontsource/merriweather';
+
+// Highlighting for code blocks
+import 'prismjs/themes/prism.css';