+++ src/components/Layout.tsx
@@ -1,0 +1,19 @@
+import * as React from 'react';
+
+import Navbar from './Navbar';
+import Footer from './Footer';
+
+const Layout = ({ location, title, children }: any): any => {
+  const rootPath = `${__PATH_PREFIX__}/`;
+  const isRootPath = location?.pathname === rootPath;
+
+  return (
+    <div className="global-wrapper" data-is-root-path={isRootPath}>
+      <Navbar />
+      <main>{children}</main>
+      <Footer />
+    </div>
+  );
+};
+
+export default Layout;