+++ src/components/Navbar.tsx
@@ -1,89 +1,74 @@
-import { Disclosure } from "@headlessui/react";
-import { MenuIcon, XIcon } from "@heroicons/react/outline";
-import NavLink from "./ui_elements/NavLink";
-import NavDisclosureButton from "./ui_elements/NavDisclosureButton";
+import React, { Fragment } from 'react';
+import { Disclosure } from '@headlessui/react';
+import { Bars3Icon, XMarkIcon } from '@heroicons/react/24/outline';
+import { Link } from 'gatsby';
 
-interface IProps {
-  navLocation: string;
-  setNavLocation: (val: string) => void;
+const navigation = [
+  { name: 'Blog', href: '' },
+  { name: 'Portfolio', href: 'portfolio' }
+];
+
+function classNames (...classes: any[]): any {
+  return classes.filter(Boolean).join(' ');
 }
 
-export default function Navbar({ navLocation, setNavLocation }: IProps) {
+export default function Example (): any {
   return (
-    <Disclosure as="nav" className="bg-white shadow border fixed w-full z-50">
+    <Disclosure as="nav" className="bg-gray-800">
       {({ open }) => (
         <>
-          <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
-            <div className="flex justify-between sm:justify-around px-12 h-16">
-              <div className="flex">
-                {/* Nav Icon */}
+          <div className="mx-auto max-w-7xl px-2 sm:px-6 lg:px-8">
+            <div className="relative flex h-16 items-center justify-between">
+              <div className="absolute inset-y-0 left-0 flex items-center sm:hidden">
+                {/* Mobile menu button */}
+                <Disclosure.Button className="inline-flex items-center justify-center rounded-md p-2 text-gray-400 hover:bg-gray-700 hover:text-white focus:outline-none focus:ring-2 focus:ring-inset focus:ring-white">
+                  <span className="sr-only">Open main menu</span>
+                  {open
+                    ? (
+                      <XMarkIcon className="block h-6 w-6" aria-hidden="true" />
+                      )
+                    : (
+                      <Bars3Icon className="block h-6 w-6" aria-hidden="true" />
+                      )}
+                </Disclosure.Button>
+              </div>
+              <div className="flex flex-1 items-center justify-center sm:items-stretch">
+                {/* Nav Icon 🐳 */}
                 <div className="flex-shrink-0 flex items-center">
-                  <a href="./">
-                    <img className="h-8 w-auto" src="./whale.png" alt="whale" />
+                  <a href="/">
+                    <img className="h-8 w-auto" src="/whale.png" alt="whale" />
                   </a>
                 </div>
-                {/* Nav Items */}
-                <div className="hidden sm:ml-6 sm:flex sm:space-x-8">
-                  <NavLink
-                    href="about"
-                    navLocation={navLocation}
-                    setNavLocation={setNavLocation}
-                  />
-                  <NavLink
-                    href="skills"
-                    navLocation={navLocation}
-                    setNavLocation={setNavLocation}
-                  />
-                  <NavLink
-                    href="projects"
-                    navLocation={navLocation}
-                    setNavLocation={setNavLocation}
-                  />
-                  <NavLink
-                    href="contact"
-                    navLocation={navLocation}
-                    setNavLocation={setNavLocation}
-                  />
+                <div className="hidden sm:ml-6 sm:block">
+                  <div className="flex space-x-4">
+                    {navigation.map(item => (
+                      <Link to={`/${item.href}`}
+                        className={classNames(
+                          'text-gray-300 hover:bg-gray-700 hover:text-white',
+                          'px-3 py-2 rounded-md text-sm font-medium'
+                        )}
+                        key={item.name}
+                      >{item.name}</Link>
+                    ))}
+                  </div>
                 </div>
-              </div>
-              {/* Responsive */}
-              <div className="-mr-2 flex items-center sm:hidden">
-                {/* 🍔 menu button */}
-                <Disclosure.Button className="inline-flex items-center justify-center p-2 rounded-md text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
-                  <span className="sr-only">Open main menu</span>
-                  {open ? (
-                    <XIcon className="block h-6 w-6" aria-hidden="true" />
-                  ) : (
-                    <MenuIcon className="block h-6 w-6" aria-hidden="true" />
-                  )}
-                </Disclosure.Button>
               </div>
             </div>
           </div>
 
-          {/* 🍔 menu options */}
           <Disclosure.Panel className="sm:hidden">
-            <div className="pt-2 pb-3 space-y-1">
-              <NavDisclosureButton
-                href="about"
-                navLocation={navLocation}
-                setNavLocation={setNavLocation}
-              ></NavDisclosureButton>
-              <NavDisclosureButton
-                href="skills"
-                navLocation={navLocation}
-                setNavLocation={setNavLocation}
-              ></NavDisclosureButton>
-              <NavDisclosureButton
-                href="projects"
-                navLocation={navLocation}
-                setNavLocation={setNavLocation}
-              ></NavDisclosureButton>
-              <NavDisclosureButton
-                href="contact"
-                navLocation={navLocation}
-                setNavLocation={setNavLocation}
-              ></NavDisclosureButton>
+            <div className="space-y-1 px-2 pt-2 pb-3 flex flex-col">
+              {navigation.map(item => (
+                <Link to={`/${item.href}`}
+                  className={classNames(
+                    'text-gray-300 hover:bg-gray-700 hover:text-white',
+                    'px-3 py-2 rounded-md text-sm font-medium'
+                  )}
+                  key={item.name}
+                >
+                  {item.name}
+                </Link>
+              ))}
             </div>
           </Disclosure.Panel>
         </>