+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+02-01-2023
+
+- change from Vite to Gatsby (for blogging)
+
 08-06-2022
 
 - change from HTML to Vite